# Conflicts:
#	docs/api.md
#	src/main/java/com/better/CommuteMate/schedule/controller/schedule/dtos/WorkScheduleHistoryResponse.java

+++ src/main/java/com/better/CommuteMate/auth/controller/dto/RegisterRequest.java
@@ -29,6 +29,6 @@ public class RegisterRequest {
     @NotNull
     private CodeType roleCode;
 
-    private Integer organizationId;
+    private Long organizationId;
 
 }
