+++ src/main/java/com/better/CommuteMate/schedule/controller/schedule/dtos/WorkScheduleHistoryResponse.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 
 @Getter
 public class WorkScheduleHistoryResponse extends ResponseDetail {
-    private final Integer id;
+    private final Long id;
     private final String userName;
     private final LocalDateTime start;
     private final LocalDateTime end;
@@ -21,7 +21,7 @@ public class WorkScheduleHistoryResponse extends ResponseDetail {
     private final Long workDurationMinutes;
 
     @Builder
-    public WorkScheduleHistoryResponse(Integer id, String userName, LocalDateTime start, LocalDateTime end, CodeType status,
+    public WorkScheduleHistoryResponse(Long id, String userName, LocalDateTime start, LocalDateTime end, CodeType status,
                                        LocalDateTime actualStart, LocalDateTime actualEnd, Long workDurationMinutes) {
         this.id = id;
         this.userName = userName;
