+++ src/main/java/com/better/CommuteMate/task/controller/dtos/AssigneeOverride.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class AssigneeOverride {
 
     private Long itemId; // 템플릿 항목 ID
-    private Integer assigneeId; // 오버라이드할 담당자 ID
+    private Long assigneeId; // 오버라이드할 담당자 ID
 }
